refactor: separate date input item
